Split some density classes into a separate header
